+++ src/ProEditor/index.ts
@@ -2,8 +2,9 @@ export { ProEditor, ProEditorProvider } from './container';
 export type { ProEditorProps, StoreUpdaterProps } from './container';
 
 // hooks
+export * from './hooks/useAssetAwareness';
 export * from './hooks/useCanvasInteraction';
-export * from './hooks/usePresenceAsset';
+export * from './hooks/useEditorAwareness';
 export * from './hooks/useProEditor';
 
 export {