+++ src/ProEditor/store/slices/general.ts
@@ -1,3 +1,4 @@
+import { StackItem } from 'yjs/dist/src/utils/UndoManager';
 import { StateCreator } from 'zustand';
 
 import getPrefixCls from '@/_util/getPrefixCls';
@@ -30,8 +31,12 @@ const initialGeneralState: GeneralSliceState = {
 
 // ======== action ======== //
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface GeneralPublicAction {}
+export interface GeneralPublicAction {
+  undo: () => void;
+  redo: () => void;
+  undoStack: () => StackItem[];
+  redoStack: () => StackItem[];
+}
 
 export interface GeneralSlice extends GeneralPublicAction, GeneralSliceState {}
 
@@ -40,6 +45,30 @@ export const generalSlice: StateCreator<
   [['zustand/devtools', never]],
   [],
   GeneralSlice
-> = () => ({
+> = (set, get) => ({
   ...initialGeneralState,
+
+  undoStack: () => {
+    return get().yjsDoc.undoManager.undoStack;
+  },
+  redoStack: () => {
+    return get().yjsDoc.undoManager.redoStack;
+  },
+  undo: () => {
+    const { yjsDoc, internalUpdateConfig } = get();
+    const stack = yjsDoc.undo();
+
+    const { config } = yjsDoc.getHistoryJSON();
+
+    internalUpdateConfig(config, { type: 'history/undo', payload: stack }, true);
+  },
+  redo: () => {
+    const { yjsDoc, internalUpdateConfig } = get();
+
+    const stack = yjsDoc.redo();
+
+    const { config } = yjsDoc.getHistoryJSON();
+
+    internalUpdateConfig(config, { type: 'history/redo', payload: stack }, true);
+  },
 });