+++ CHANGELOG.md
@@ -1,5 +1,222 @@
 # Changelog
 
+## [Version&nbsp;0.3.0-alpha.8](https://github.com/ant-design/pro-editor/compare/v0.3.0-alpha.7...v0.3.0-alpha.8)
+
+<sup>Released on **2023-06-14**</sup>
+
+#### ✨ 新特性
+
+- 完善 AssetStore 数据链路.
+
+#### 💄 样式
+
+- **leva-panel**: 添加 style 与 className 属性.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- 完善 AssetStore 数据链路 ([20df5b2](https://github.com/ant-design/pro-editor/commit/20df5b2))
+
+#### Styles
+
+- **leva-panel**: 添加 style 与 className 属性 ([b0920fe](https://github.com/ant-design/pro-editor/commit/b0920fe))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;0.3.0-alpha.7](https://github.com/ant-design/pro-editor/compare/v0.3.0-alpha.6...v0.3.0-alpha.7)
+
+<sup>Released on **2023-06-14**</sup>
+
+#### ✨ 新特性
+
+- 实现 AssetStore 与 EditorStore 的状态自动同步, 新增 LevaPanel 属性面板.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- 实现 AssetStore 与 EditorStore 的状态自动同步 ([3eeb73b](https://github.com/ant-design/pro-editor/commit/3eeb73b))
+- 新增 LevaPanel 属性面板 ([ee2d922](https://github.com/ant-design/pro-editor/commit/ee2d922))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;0.3.0-alpha.6](https://github.com/ant-design/pro-editor/compare/v0.3.0-alpha.5...v0.3.0-alpha.6)
+
+<sup>Released on **2023-06-13**</sup>
+
+#### ✨ 新特性
+
+- 重构优化 ComponentAsset store 实现，收敛 AssetStore 的定义方式.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- 重构优化 ComponentAsset store 实现，收敛 AssetStore 的定义方式 ([37e85ab](https://github.com/ant-design/pro-editor/commit/37e85ab))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;0.3.0-alpha.5](https://github.com/ant-design/pro-editor/compare/v0.3.0-alpha.4...v0.3.0-alpha.5)
+
+<sup>Released on **2023-06-13**</sup>
+
+#### 🐛 修复
+
+- 修正潜藏的 undo/redo bug，并补充单测.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- 修正潜藏的 undo/redo bug，并补充单测 ([a671b65](https://github.com/ant-design/pro-editor/commit/a671b65))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;0.3.0-alpha.4](https://github.com/ant-design/pro-editor/compare/v0.3.0-alpha.3...v0.3.0-alpha.4)
+
+<sup>Released on **2023-06-13**</sup>
+
+#### 🐛 修复
+
+- Fix config update.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- Fix config update ([fee5441](https://github.com/ant-design/pro-editor/commit/fee5441))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;0.3.0-alpha.3](https://github.com/ant-design/pro-editor/compare/v0.3.0-alpha.2...v0.3.0-alpha.3)
+
+<sup>Released on **2023-06-13**</sup>
+
+#### 🐛 修复
+
+- Remove lucide icon.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's fixed
+
+- Remove lucide icon ([697a8f3](https://github.com/ant-design/pro-editor/commit/697a8f3))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;0.3.0-alpha.2](https://github.com/ant-design/pro-editor/compare/v0.3.0-alpha.1...v0.3.0-alpha.2)
+
+<sup>Released on **2023-06-13**</sup>
+
+#### ✨ 新特性
+
+- 新增 ContextMenu 组件.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- 新增 ContextMenu 组件 ([265b442](https://github.com/ant-design/pro-editor/commit/265b442))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;0.3.0-alpha.1](https://github.com/ant-design/pro-editor/compare/v0.2.3...v0.3.0-alpha.1)
+
+<sup>Released on **2023-06-13**</sup>
+
+#### ✨ 新特性
+
+- Support undo/redo for ProEditor.
+
+#### 💄 样式
+
+- Update empty style.
+
+<br/>
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+#### What's improved
+
+- Support undo/redo for ProEditor ([cec5611](https://github.com/ant-design/pro-editor/commit/cec5611))
+
+#### Styles
+
+- Update empty style ([a6a61c5](https://github.com/ant-design/pro-editor/commit/a6a61c5))
+
+</details>
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
 ### [Version&nbsp;0.2.3](https://github.com/ant-design/pro-editor/compare/v0.2.2...v0.2.3)
 
 <sup>Released on **2023-06-13**</sup>